Create and subclass GenreDialog
The dialog box used to add genres to the database is nearly identical to the dialog to be used for editing genres already in the database. They differ mostly only in what happens when we press the 'OK' buton. Therefore, create a superclass from which these two classes can inherit their shared behaviour.